refactor(vtt): fold stat_long_filename into stat-many
Two commands stat generated names to be tracked; one varied the name's
length and the other how many names. They are now `stat-many <count>
[name-length]`, which also puts the name in kebab case with every other
subcommand.

Count leads because it is the knob that travels. A long name only fills a
channel on unix: on Windows a path reaches the tracer through a
`UNICODE_STRING` whose length is a `u16`, so no single record there exceeds
64 KiB however long a name the caller asks for.

Names now carry their index, so a run of them cannot collapse into one
record, and padding fills out whatever length is asked for. The `/dev/shm`
case keeps its one 1 MiB name as `stat-many 1 1048576`, and gains the
trailing line that reports the process survived its accesses.
