+++ crates/vt_bin/src/vtt/main.rs
@@ -26,7 +26,6 @@ mod rm;
 #[cfg(target_os = "linux")]
 mod small_dev_shm;
 mod stat_file;
-mod stat_long_filename;
 mod stat_many;
 mod touch_file;
 mod write_file;
@@ -36,7 +35,7 @@ fn main() {
     if args.len() < 2 {
         eprintln!("Usage: vtt <subcommand> [args...]");
         eprintln!(
-            "Subcommands: barrier, check-tty, cp, exit, exit-on-ctrlc, grep-file, list-dir, mkdir, pipe-stdin, print, print-color, print-cwd, print-env, print-file, read-stdin, replace-file-content, rm, small_dev_shm, stat-file, stat_long_filename, stat-many, touch-file, write-file"
+            "Subcommands: barrier, check-tty, cp, exit, exit-on-ctrlc, grep-file, list-dir, mkdir, pipe-stdin, print, print-color, print-cwd, print-env, print-file, read-stdin, replace-file-content, rm, small_dev_shm, stat-file, stat-many, touch-file, write-file"
         );
         std::process::exit(1);
     }
@@ -76,7 +75,6 @@ fn main() {
             stat_file::run(&args[2..]);
             Ok(())
         }
-        "stat_long_filename" => stat_long_filename::run(&args[2..]),
         "stat-many" => stat_many::run(&args[2..]),
         "touch-file" => touch_file::run(&args[2..]),
         "write-file" => write_file::run(&args[2..]),
